Moved resource and endpoint to endpoint.h and updated style.
